refactor: restructure temporal module and clean up related tests
- Reorganized the temporal module and relocated relevant test cases to align
  with the updated folder structure for improved maintainability.
- Moved the timestamp validation and helper functions into their appropriate
  test modules and added missing utility tests to ensure better coverage.
- Refactored `conftest.py` to encapsulate reusable test fixtures for timestamp-related tests.

Issue TCT-52
